Define a generic "Card" message type for formatted messages.
Chat interfaces which support structured output can use this to provide a better UX for hype peeps.

This also refactors how hypecore converts command responses into MessageList objects to be sent to interfaces. If a card is specified, the conversion will attempt to produce a text-only output to describe the card.

PiperOrigin-RevId: 269368242