Adds "Badges" to HypeCoffee.
Badges definitions are kept in a textproto file and loaded on startup. Functions implementing the conditionals are registered to the badge's ID and looked up at runtime. I debated on using a stored DSL, but though this was more flexible and readable at the cost of being slightly harder to implement for new badges.

PiperOrigin-RevId: 312144969